Completed feature 3: USer can save workout session data
