+++ client/pages/workout-session.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import Workout from './workout';
 import ModalButton from './modal-button';
-import Timer from './timer';
 import FinishButton from './finish-button';
 
 export default class WorkoutSession extends React.Component {
@@ -9,13 +8,28 @@ export default class WorkoutSession extends React.Component {
     super(props);
     this.state = {
       workouts: [],
-      sessionDuration: 0
+
+      seconds: 0,
+      timerOn: true,
+      timerId: null
     };
 
     this.addWorkout = this.addWorkout.bind(this);
     this.handleWorkoutUpdated = this.handleWorkoutUpdated.bind(this);
-    this.addSessionDuration = this.addSessionDuration.bind(this);
+
     this.postRequest = this.postRequest.bind(this);
+    this.click = this.click.bind(this);
+    this.renderPlayOrPause = this.renderPlayOrPause.bind(this);
+
+  }
+
+  componentDidMount() {
+    const mountTimerId = setInterval(() => {
+      this.setState({ seconds: this.state.seconds + 1 });
+
+    }, 1000);
+
+    this.setState({ timerId: mountTimerId });
 
   }
 
@@ -39,14 +53,8 @@ export default class WorkoutSession extends React.Component {
     this.setState({ workouts: updatedWorkouts });
   }
 
-  addSessionDuration(seconds) {
-    const minutes = Math.floor(seconds / 60);
-    this.setState({ sessionDuration: minutes });
-
-  }
-
   postRequest(name) {
-    const { workouts, sessionDuration } = this.state;
+    const { workouts, seconds } = this.state;
     if (this.state.workouts.length < 1) {
       return;
     }
@@ -54,32 +62,33 @@ export default class WorkoutSession extends React.Component {
 
       session: {
         name: name,
-        durationInMinutes: sessionDuration,
+        durationInMinutes: Math.floor(seconds / 60),
         userId: 1
       },
       sessionWorkouts: []
 
     };
-    let counter = 0;
+    let objectInArray = {};
+
     for (let i = 0; i < workouts.length; i++) {
       for (let j = 0; j < workouts[i].sets.length; j++) {
-        submitData.sessionWorkouts.push(workouts[i].sets[j]);
-        submitData.sessionWorkouts[counter].set = j + 1;
-        if (submitData.sessionWorkouts[counter].reps < 0 || submitData.sessionWorkouts[counter].weight < 0) {
+        objectInArray.reps = workouts[i].sets[j].reps;
+        objectInArray.weight = workouts[i].sets[j].weight;
+        objectInArray.workoutId = workouts[i].sets[j].workoutId;
+        if (parseInt(objectInArray.reps) < 0 || parseInt(objectInArray.weight) < 0) {
           window.alert('Cannot have values that are negative. Please fix and try again.');
           return;
         }
 
-        if (submitData.sessionWorkouts[counter].reps === '' ||
-           submitData.sessionWorkouts[counter].weight === '') {
+        if (objectInArray.reps === '' ||
+          objectInArray.weight === '') {
           window.alert('Cannot have values that are blank. Please add numerical values.');
           return;
         }
+        submitData.sessionWorkouts.push(objectInArray);
 
-        counter++;
-
+        objectInArray = {};
       }
-
     }
 
     fetch('api/add-session-and-session-workout-data',
@@ -94,10 +103,67 @@ export default class WorkoutSession extends React.Component {
       });
 
     this.setState({ workouts: [] });
+    clearInterval(this.state.timerId);
+    this.setState({ timerOn: !this.state.timerId });
+    this.setState({ seconds: 0 });
+
+    const mountTimerId = setInterval(() => {
+      this.setState({ seconds: this.state.seconds + 1 });
+
+    }, 1000);
+
+    this.setState({ timerId: mountTimerId });
+    this.setState({ timerOn: !this.state.timerOn });
 
   }
 
+  click() {
+
+    if (this.state.timerOn) {
+      clearInterval(this.state.timerId);
+      this.setState({ timerOn: !this.state.timerOn });
+    } else {
+
+      const mountTimerId = setInterval(() => {
+        this.setState({ seconds: this.state.seconds + 1 });
+
+      }, 1000);
+
+      this.setState({ timerId: mountTimerId });
+      this.setState({ timerOn: !this.state.timerOn });
+
+    }
+
+  }
+
+  renderPlayOrPause() {
+    if (this.state.timerOn) {
+      return (
+        <i onClick={this.click} className="bi bi-pause-circle me-5 h1"></i>
+      );
+    } else {
+      return (
+        <i onClick={this.click} className="bi bi-play-circle me-5 h1"></i>
+      );
+    }
+  }
+
   render() {
+    let seconds = 0;
+    let minutes = 0;
+    let hours = 0;
+    hours = Math.floor(this.state.seconds / 60 / 60);
+    seconds = 60 * hours * 60;
+    seconds = this.state.seconds - seconds;
+    minutes = Math.floor(seconds / 60);
+    seconds = (60 * minutes) + (60 * 60 * hours);
+    let leftOverSeconds = this.state.seconds - seconds;
+    if (minutes < 10) {
+      minutes = `0${minutes}`;
+    }
+    if (leftOverSeconds < 10) {
+      leftOverSeconds = `0${leftOverSeconds}`;
+    }
 
     return (
     <>
@@ -109,7 +175,10 @@ export default class WorkoutSession extends React.Component {
 
             <div className='col-6'>
 
-              <Timer addSessionDuration={this.addSessionDuration} />
+              <div className='d-flex no-wrap justify-content-end'>
+                {this.renderPlayOrPause()}
+                <p className='text-end'>{`${hours}:${minutes}:${leftOverSeconds}`}</p>
+              </div>
               <div className='d-flex justify-content-end'>
                   <FinishButton postRequest={this.postRequest} />
                   <button type='button' className='btn btn-primary btn-md ms-4'>Cancel</button>
